LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-input-and-change-events-expected.txt

Test for keyboard operations for <input type=datetime-local>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 1
PASS changeEventsFired is 0
PASS input.value is "2012-09-02T00:02"
PASS inputEventsFired is 1
PASS changeEventsFired is 1
PASS successfullyParsed is true

TEST COMPLETE
